64 bits de endereçamento

+++ include/architecture/armv8/armv8_mmu.h
@@ -18,10 +18,10 @@ private:
 
     static const bool colorful = Traits<MMU>::colorful;
     static const unsigned int COLORS = Traits<MMU>::COLORS;
-    static const unsigned int RAM_BASE = Memory_Map::RAM_BASE;
-    static const unsigned int APP_LOW = Memory_Map::APP_LOW;
-    static const unsigned int PHY_MEM = Memory_Map::PHY_MEM;
-    static const unsigned int SYS = Memory_Map::SYS;
+    static const unsigned long RAM_BASE = Memory_Map::RAM_BASE;
+    static const unsigned long APP_LOW = Memory_Map::APP_LOW;
+    static const unsigned long PHY_MEM = Memory_Map::PHY_MEM;
+    static const unsigned long SYS = Memory_Map::SYS;
 
 public:
     // Page Flags
